backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose==3.3.0
pydantic==2.7.0
